# app/Virtual/Resources/CompanyResource.php
<?php


namespace App\Virtual\Resources;


/**
 * @OA\Schema(
 *     title="CompanyResource",
 *     description="Company resource",
 *     @OA\Xml(
 *         name="CompanyResource"
 *     )
 * )
 */
class CompanyResource
{
    /**
     * @OA\Property(
     *     title="Data",
     *     description="Data wrapper"
     * )
     *
     * @var \App\Virtual\Models\Company
     */
    private $data;
}