# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gigs API",
        "description": "API documentation for Gigs API.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/company": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "GET (all) companies",
                "description": "Returns list of companies for auth user",
                "operationId": "GetCompanies",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Company"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "POST (add) company",
                "description": "Returns company",
                "operationId": "addCompany",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/company/{id}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "GET Company",
                "description": "Show company",
                "operationId": "showCompany",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Company"
                ],
                "summary": "UPDATE Company",
                "description": "UPDATE Company",
                "operationId": "updateCompany",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Company"
                ],
                "summary": "DELETE Company",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteCompany",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The resource was deleted successfully."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "GET user profile",
                "description": "Profile",
                "operationId": "profile",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "GET User",
                "description": "Show user",
                "operationId": "showUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "UPDATE User",
                "description": "UPDATE user",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CompanyRequest": {
                "title": "Create Company Request",
                "description": "Create Company request body data",
                "required": [
                    "name",
                    "address"
                ],
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Company's name",
                        "type": "string",
                        "example": "Company name"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the company",
                        "type": "string",
                        "example": "Doe"
                    },
                    "address": {
                        "title": "Address",
                        "description": "Company's address",
                        "type": "string",
                        "example": "Company's address"
                    }
                },
                "type": "object"
            },
            "CompanyUpdateRequest": {
                "title": "Update Company Request",
                "description": "Update Company request body data",
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Company's name",
                        "type": "string",
                        "example": "Company name"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the company",
                        "type": "string",
                        "example": "Doe"
                    },
                    "address": {
                        "title": "Address",
                        "description": "Company's address",
                        "type": "string",
                        "example": "Company's address"
                    }
                },
                "type": "object"
            },
...
            "Company": {
                "title": "Company",
                "description": "Company model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Company's name",
                        "type": "string",
                        "example": "Company name"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the company",
                        "type": "string",
                        "example": "Doe"
                    },
                    "address": {
                        "title": "Address",
                        "description": "Company's address",
                        "type": "string",
                        "example": "Company's address"
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 10
                            },
                            "first_name": {
                                "type": "string",
                                "example": "John"
                            },
                            "last_name": {
                                "type": "string",
                                "example": "John"
                            },
                            "email": {
                                "type": "string",
                                "example": "johndoe@example.com"
                            }
                        },
                        "type": "object"
                    },
                    "gigs": {
                        "title": "user gigs",
                        "description": "user gigs",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "name": {
                                    "type": "string",
                                    "example": "gig name"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "gig description"
                                },
                                "timestamp_start": {
                                    "type": "string",
                                    "example": "09/24/2024 12:00 PM"
                                },
                                "timestamp_end": {
                                    "type": "string",
                                    "example": "09/25/2024 12:00 PM"
                                },
                                "number_of_positions": {
                                    "type": "integer",
                                    "example": 45
                                },
                                "pay_per_hour": {
                                    "type": "number",
                                    "format": "double",
                                    "example": 4.5
                                },
                                "status": {
                                    "type": "boolean",
                                    "example": 0
                                },
                                "company": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "company_name"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Company"
                }
            },
...
            "CompanyResource": {
                "title": "CompanyResource",
                "description": "Company resource",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Company"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CompanyResource"
                }
            },
...
            "UserResource": {
                "title": "UserResource",
                "description": "User resource",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserResource"
                }
            },
...
            "UserUpdateRequest": {
                "title": "Update User Request",
                "description": "Update User request body data",
                "properties": {
                    "first_name": {
                        "title": "First name",
                        "description": "User's first name",
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "title": "Last Name",
                        "description": "User's last name",
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "title": "Email",
                        "description": "User's email address",
                        "type": "string",
                        "example": "john@doe.com"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password must be at least 8 characters long and contain a mix of letters and numbers.",
                        "type": "string",
                        "example": "passwordCreate12"
                    },
                    "password_confirmation": {
                        "title": "Password confirmation",
                        "description": "Password confirmation",
                        "type": "string",
                        "example": "passwordCreate12"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://gigs.devs/oauth/authorize",
                        "tokenUrl": "http://gigs.devs/oauth/token",
                        "refreshUrl": "http://gigs.devs/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}